implement product ownership and permissions + move validation to products.js
